requirements.txt
Django>=3.2,<4.0
djangorestframework
drf-yasg
pyodbc
mssql-django
